Switch to "clean" hall geometry
Original file seemed to override some of the positions or rotations or
solids or volumes from arc2x2.gdml, causing checkOverlaps.C to report a
couple of overlaps in the cryostat. There are no overlaps in arc2x2.gdml
by itself. You'd think that the "parent" file shouldn't be able to
override things in the "child" file but I guess not. So the solution is
to just remove everything cryostat-related in the hall GDML.